remove unneccessary cast

+++ src/embbeded_datasets.py
@@ -32,7 +32,7 @@ class embbededDataset:
         train_codes = torch.Tensor(torch.load(os.path.join(self.dataset_loc, "train_codes.pt")))
         labels = torch.load(os.path.join(self.dataset_loc, "train_labels.pt")).float()
 
-        train_labels = torch.Tensor(labels).cpu().float()
+        train_labels = torch.Tensor(labels).cpu()
 
         if self.transform:
             if self.transform == "standard":
